cmd/hiveview: improve esbuild setup (#710)
In #708, I introduced a lot of logic to create CSS/JS bundles on demand. It turns out that
the way I implemented bundling was a bit misguided, because I was operating under the
basic assumption that esbuild will always create exactly one output file for every entrypoint.
This breaks when it tries to copy assets or performs code splitting. So, instead of doing
incremental builds ourself, just let esbuild do its thing and copy all of the files it produces
into the output FS.

As a bonus, enabling code splitting makes it possible to have one entry point for each page.